+++ src/main/java/repository/article/ArticleRepository.java
@@ -1,4 +1,4 @@
-package repository;
+package repository.article;
 
 import domain.Article;
